Split out BrowseList and BrowseItem into separate files.
