+++ browse-handler/java/org/vufind/solr/handler/BrowseRequestHandler.java
@@ -496,72 +496,6 @@ class BibDB
 
 
 
-class BrowseList
-{
-    public int totalCount;
-    public List<BrowseItem> items = new ArrayList<> ();
-
-
-    public List<Map<String, Object>> asMap()
-    {
-        List<Map<String, Object>> result = new ArrayList<> ();
-
-        for (BrowseItem item : items) {
-            result.add(item.asMap());
-        }
-
-        return result;
-    }
-}
-
-
-
-class BrowseItem
-{
-    public List<String> seeAlso = new ArrayList<> ();
-    public List<String> useInstead = new ArrayList<> ();
-    public String note = "";
-    public String sort_key;
-    public String heading;
-    public List<String> ids;
-    public Map<String, List<Collection<String>>> extras = new HashMap<> ();
-    int count;
-
-
-    public BrowseItem(String sort_key, String heading)
-    {
-        this.sort_key = sort_key;
-        this.heading = heading;
-    }
-
-    // ids are gathered into List<Collection<String>>, see bibinfo in
-    // BibDB.matchingIDs() and populateItem().
-    public void setIds(List<Collection<String>> idList)
-    {
-        ids = new ArrayList<String> ();
-        for (Collection<String> idCol : idList) {
-            ids.addAll(idCol);
-        }
-        this.ids = ids;
-    }
-
-    public Map<String, Object> asMap()
-    {
-        Map<String, Object> result = new HashMap<> ();
-
-        result.put("sort_key", sort_key);
-        result.put("heading", heading);
-        result.put("seeAlso", seeAlso);
-        result.put("useInstead", useInstead);
-        result.put("note", note);
-        result.put("count", new Integer(count));
-        result.put("ids", ids);
-        result.put("extras", extras);
-
-        return result;
-    }
-}
-
 
 
 
@@ -638,7 +572,6 @@ class Browse
         return result;
     }
 }
-
 
 
 class BrowseSource
